Split out Author
